Allow enforcing namespace to HTTP solvers
Currently cert-manager uses the namespace where the ingress with tls
annotation object exist to create all necessary resources for the http
acme challenge (pod, service and ingress).

This has the following drawbacks:

* Requires cert-manager to have CREATE permissions in all namespaces
* Because resourceQuotas, some namespaces will not allow to create the
pod for the acme challenge
* The customer in a namespace will see the pods of the challenge on
their namespace. We want to hide that logic from them.

This commit introduces a new option (ACMEHTTP01SolverNamespace) and a
property to ACMEOptions that will be used when building a challenge
solver pod spec.  If a value is set for said property, the pod will
respect that namespace. Otherwise it'll take the one proposed by the
Challenge object.

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>